# Conflicts:
#	src/main/java/com/ghostchu/peerbanhelper/gui/PBHGuiBridge.java

+++ src/main/java/com/ghostchu/peerbanhelper/databasent/service/AlertService.java
@@ -14,7 +14,7 @@ public interface AlertService extends IService<AlertEntity> {
 
     boolean identifierAlertExistsIncludeRead(@NotNull String identifier);
 
-    int deleteOldAlerts(@NotNull OffsetDateTime before);
+    long deleteOldAlerts(@NotNull OffsetDateTime before);
 
     void markAllAsRead();
